Develop v1

+++ DU_AN_1/src/main/java/views/FrmQLPhuPhi.java
@@ -4,6 +4,7 @@ import enumclass.trangThaiPhuPhi;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
 import modelview.QLPhuPhi;
 import service.IPhuPhiService;
@@ -18,27 +19,53 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
     public FrmQLPhuPhi() {
         initComponents();
         showDataTable();
+        txtMa.setEnabled(false);
+    }
+
+    private String checkTrong() {
+        if (txtTen.getText().isBlank()) {
+            return "KHÔNG ĐƯỢC BỎ TRỐNG";
+        } else if (txtTen.getText().matches("^[0-9]+$")) {
+            return "TÊN CÓ SỐ RỒI";
+        } else {
+            return "OK";
+        }
     }
 
     private QLPhuPhi formInputqLPhuPhiSave() {
+        String check = checkTrong();
         QLPhuPhi ql = new QLPhuPhi();
-        ql.setId(null);
-        String ma = txtMa.getText();
-        ql.setMaPhuPhi(ma);
-        String ten = txtTen.getText();
-        ql.setTenPhuPhi(ten);
-        return ql;
+        if (check != "OK") {
+            JOptionPane.showMessageDialog(this, check);
+        } else {
+            ql.setId(null);
+            String ma = iPhuPhiService.genMaPhuPhi(lstQLPhuPhis);
+            ql.setMaPhuPhi(ma);
+            String ten = txtTen.getText();
+            ql.setTenPhuPhi(ten);
+            return ql;
+        }
+        return null;
     }
 
     private QLPhuPhi formInputqLPhuPhiUpdate() {
+        String check = checkTrong();
         QLPhuPhi ql = new QLPhuPhi();
-        String id = txtID.getText();
-        ql.setId(id);
-        String ma = txtMa.getText();
-        ql.setMaPhuPhi(ma);
-        String ten = txtTen.getText();
-        ql.setTenPhuPhi(ten);
-        return ql;
+        if (check != "OK") {
+            JOptionPane.showMessageDialog(this, check);
+        } else {
+            int row = tbQLPhuPhi.getSelectedRow();
+            QLPhuPhi qlgetID = lstQLPhuPhis.get(row);
+            String id = qlgetID.getId();
+            ql.setId(id);
+            String ma = txtMa.getText();
+            ql.setMaPhuPhi(ma);
+            String ten = txtTen.getText();
+            ql.setTenPhuPhi(ten);
+            return ql;
+        }
+        return null;
+
     }
 
     private void showDataTable() {
@@ -46,7 +73,7 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
         lstQLPhuPhis = new ArrayList<>();
         iPhuPhiService = new PhuPhiServiceImpl();
         lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
-        String[] header = {"ID", "MÃ", "TÊN"};
+        String[] header = {"MÃ", "TÊN"};
         dtm.setColumnIdentifiers(header);
         loadDataTable(lstQLPhuPhis);
     }
@@ -54,7 +81,7 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
     private void loadDataTable(List<QLPhuPhi> lstQLPhuPhis) {
         dtm.setRowCount(0);
         for (QLPhuPhi l : lstQLPhuPhis) {
-            Object[] toData = {l.getId(), l.getMaPhuPhi(), l.getTenPhuPhi()};
+            Object[] toData = {l.getMaPhuPhi(), l.getTenPhuPhi()};
             dtm.addRow(toData);
         }
     }
@@ -67,18 +94,19 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
         jLabel3 = new javax.swing.JLabel();
         btnThem = new javax.swing.JButton();
         btnSua = new javax.swing.JButton();
-        txtID = new javax.swing.JTextField();
         btnXoa = new javax.swing.JButton();
         txtMa = new javax.swing.JTextField();
         jSeparator1 = new javax.swing.JSeparator();
         jLabel1 = new javax.swing.JLabel();
         jSeparator2 = new javax.swing.JSeparator();
-        jLabel2 = new javax.swing.JLabel();
         jScrollPane1 = new javax.swing.JScrollPane();
         tbQLPhuPhi = new javax.swing.JTable();
         jSeparator3 = new javax.swing.JSeparator();
         txtTen = new javax.swing.JTextField();
         jLabel4 = new javax.swing.JLabel();
+        txtTimKiem = new javax.swing.JTextField();
+        jLabel5 = new javax.swing.JLabel();
+        btnTimKiem = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
@@ -108,8 +136,6 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
         jLabel1.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
         jLabel1.setText("QUẢN LÝ PHỤ PHÍ");
 
-        jLabel2.setText("ID:");
-
         tbQLPhuPhi.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
                 {null, null, null, null},
@@ -130,6 +156,15 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
 
         jLabel4.setText("TÊN:");
 
+        jLabel5.setText("TÌM KIẾM:");
+
+        btnTimKiem.setText("TÌM KIẾM");
+        btnTimKiem.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnTimKiemActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
@@ -137,15 +172,14 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
             .addGroup(layout.createSequentialGroup()
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(layout.createSequentialGroup()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                            .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addGroup(javax.swing.GroupLayout.Alignment.LEADING, layout.createSequentialGroup()
                                 .addGap(266, 266, 266)
-                                .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 288, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(311, 311, 311))
+                                .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 288, javax.swing.GroupLayout.PREFERRED_SIZE))
                             .addGroup(javax.swing.GroupLayout.Alignment.LEADING, layout.createSequentialGroup()
                                 .addContainerGap()
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                                    .addComponent(jScrollPane1, javax.swing.GroupLayout.Alignment.LEADING)
+                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                                    .addComponent(jScrollPane1, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, 954, Short.MAX_VALUE)
                                     .addComponent(jSeparator2, javax.swing.GroupLayout.Alignment.LEADING))))
                         .addGap(0, 0, Short.MAX_VALUE))
                     .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
@@ -160,52 +194,65 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
                         .addGap(0, 0, Short.MAX_VALUE))
                     .addGroup(layout.createSequentialGroup()
                         .addComponent(btnThem, javax.swing.GroupLayout.PREFERRED_SIZE, 219, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addGap(31, 31, 31)
                         .addComponent(btnSua, javax.swing.GroupLayout.PREFERRED_SIZE, 219, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(97, 97, 97)
+                        .addGap(18, 18, 18)
                         .addComponent(btnXoa, javax.swing.GroupLayout.PREFERRED_SIZE, 219, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 30, Short.MAX_VALUE)
+                        .addComponent(btnTimKiem, javax.swing.GroupLayout.PREFERRED_SIZE, 219, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addContainerGap())
                     .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addGap(69, 69, 69)
-                        .addComponent(jLabel2)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(txtID, javax.swing.GroupLayout.PREFERRED_SIZE, 333, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                         .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                             .addGroup(layout.createSequentialGroup()
-                                .addComponent(jLabel4)
+                                .addGap(57, 57, 57)
+                                .addComponent(jLabel3)
+                                .addGap(21, 21, 21)
+                                .addComponent(txtMa, javax.swing.GroupLayout.PREFERRED_SIZE, 297, javax.swing.GroupLayout.PREFERRED_SIZE)
                                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addComponent(jLabel4)
+                                .addGap(30, 30, 30)
                                 .addComponent(txtTen, javax.swing.GroupLayout.PREFERRED_SIZE, 297, javax.swing.GroupLayout.PREFERRED_SIZE))
                             .addGroup(layout.createSequentialGroup()
-                                .addComponent(jLabel3)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addComponent(txtMa, javax.swing.GroupLayout.PREFERRED_SIZE, 297, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addGap(0, 0, Short.MAX_VALUE)
+                                .addComponent(jLabel5)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                .addComponent(txtTimKiem, javax.swing.GroupLayout.PREFERRED_SIZE, 297, javax.swing.GroupLayout.PREFERRED_SIZE)))
                         .addGap(42, 42, 42))))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(jLabel1)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel1)
+                    .addComponent(jLabel5)
+                    .addComponent(txtTimKiem, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jSeparator1, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(9, 9, 9)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel2)
-                    .addComponent(jLabel3)
-                    .addComponent(txtID, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(txtMa, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(17, 17, 17)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel4)
-                    .addComponent(txtTen, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(12, 12, 12)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(txtTen, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel4)))
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(txtMa, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel3))))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jSeparator3, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(7, 7, 7)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnThem)
-                    .addComponent(btnSua)
-                    .addComponent(btnXoa))
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(7, 7, 7)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(btnThem)
+                            .addComponent(btnSua)))
+                    .addGroup(layout.createSequentialGroup()
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(btnTimKiem)
+                            .addComponent(btnXoa))))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jSeparator2, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
@@ -217,37 +264,78 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
 
     private void btnThemActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnThemActionPerformed
         QLPhuPhi ql = formInputqLPhuPhiSave();
-        iPhuPhiService.save(ql);
-        lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
-        loadDataTable(lstQLPhuPhis);
+        if (ql == null) {
+            return;
+        } else {
+            iPhuPhiService.save(ql);
+            JOptionPane.showMessageDialog(this, "THÀNH CÔNG");
+            lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
+            loadDataTable(lstQLPhuPhis);
+        }
     }//GEN-LAST:event_btnThemActionPerformed
 
     private void btnSuaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSuaActionPerformed
         // TODO add your handling code here:
         QLPhuPhi ql = formInputqLPhuPhiUpdate();
-        iPhuPhiService.update(ql);
-        lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
-        loadDataTable(lstQLPhuPhis);
+        if (ql == null) {
+            return;
+        } else {
+            iPhuPhiService.update(ql);
+            JOptionPane.showMessageDialog(this, "THÀNH CÔNG");
+            lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
+            loadDataTable(lstQLPhuPhis);
+        }
     }//GEN-LAST:event_btnSuaActionPerformed
 
     private void btnXoaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnXoaActionPerformed
         // TODO add your handling code here:
-        int row = tbQLPhuPhi.getSelectedRow();
-        QLPhuPhi qLPhuPhi = lstQLPhuPhis.get(row);
-        String idQLPhuPhi = qLPhuPhi.getId();
-        iPhuPhiService.delete(idQLPhuPhi);
-        lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
-        loadDataTable(lstQLPhuPhis);
+        if (!txtMa.getText().isEmpty()) {
+            int row = tbQLPhuPhi.getSelectedRow();
+            QLPhuPhi qLPhuPhi = lstQLPhuPhis.get(row);
+            String idQLPhuPhi = qLPhuPhi.getId();
+            iPhuPhiService.delete(idQLPhuPhi);
+            JOptionPane.showMessageDialog(this, "THÀNH CÔNG");
+            lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
+            loadDataTable(lstQLPhuPhis);
+        } else {
+            JOptionPane.showMessageDialog(this, "CHƯA CHỌN");
+            return;
+        }
+
     }//GEN-LAST:event_btnXoaActionPerformed
 
     private void tbQLPhuPhiMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tbQLPhuPhiMouseClicked
         // TODO add your handling code here:
         int row = tbQLPhuPhi.getSelectedRow();
         QLPhuPhi ql = lstQLPhuPhis.get(row);
-        txtID.setText(String.valueOf(ql.getId()));
         txtMa.setText(ql.getMaPhuPhi());
         txtTen.setText(ql.getTenPhuPhi());
     }//GEN-LAST:event_tbQLPhuPhiMouseClicked
+
+    private void btnTimKiemActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnTimKiemActionPerformed
+        // TODO add your handling code here:
+        lstQLPhuPhis = iPhuPhiService.getAllQLPhuPhis();
+        String timKiem = txtTimKiem.getText();
+        if (timKiem.isBlank()) {
+            JOptionPane.showMessageDialog(this, "TÌM KIẾM TRÔNG");
+            return;
+        }
+        int dem = 0;
+        for (QLPhuPhi lst : lstQLPhuPhis) {
+            if (lst.getMaPhuPhi().equals(timKiem)) {
+                dem = dem + 1;
+            }
+        }
+        if (dem == 1) {
+            lstQLPhuPhis.clear();
+            QLPhuPhi qLPhuPhi = iPhuPhiService.fillByMaPhuPhi(timKiem);
+            lstQLPhuPhis.add(qLPhuPhi);
+            loadDataTable(lstQLPhuPhis);
+        } else {
+            JOptionPane.showMessageDialog(this, "KHÔNG CÓ");
+            return;
+        }
+    }//GEN-LAST:event_btnTimKiemActionPerformed
 
     /**
      * @param args the command line arguments
@@ -296,19 +384,20 @@ public class FrmQLPhuPhi extends javax.swing.JFrame {
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton btnSua;
     private javax.swing.JButton btnThem;
+    private javax.swing.JButton btnTimKiem;
     private javax.swing.JButton btnXoa;
     private javax.swing.ButtonGroup buttonGroup1;
     private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
     private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JSeparator jSeparator1;
     private javax.swing.JSeparator jSeparator2;
     private javax.swing.JSeparator jSeparator3;
     private javax.swing.JTable tbQLPhuPhi;
-    private javax.swing.JTextField txtID;
     private javax.swing.JTextField txtMa;
     private javax.swing.JTextField txtTen;
+    private javax.swing.JTextField txtTimKiem;
     // End of variables declaration//GEN-END:variables
 }
